+++ app/api/cards/route.js
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import dbConnect from "@/lib/mongo/dbConnect";
+import { getSession } from "next-auth/react";
 import Card from "@/models/Card";
 
 /**
@@ -20,8 +21,15 @@ export async function GET(req, res) {
 
 export async function POST(req) {
   await dbConnect();
-
+  const session = await getSession({ req });
+  if (!session) {
+    return NextResponse.json({ success: false, message: "Unauthorized" }, { status: 401 });
+  }
+  const userId = session.user.id;
+  console.log(req);
   const body = await req.json();
+  console.log(body);
+  body.createdBy = userId;
   console.log(body);
   // Create a new card based on the request body
   const card = await Card.create(body);
@@ -33,8 +41,12 @@ export async function POST(req) {
 
 export async function DELETE(req) {
   await dbConnect();
-
-  const cards = await Card.deleteMany({});
+  const session = await getSession({ req });
+  if (!session) {
+    return NextResponse.json({ success: false, message: "Unauthorized" }, { status: 401 });
+  }
+  const userId = session.user.id;
+  const cards = await Card.deleteMany({ createdBy: userId });
   if (!cards) {
     return NextResponse.json({ success: false, message: "No cards found" }, { status: 400 });
   }
